+++ srcs/string/ft_strrchr.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strrchr.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tkomatsu <tkomatsu@student.42tokyo.jp>     +#+  +:+       +#+        */
+/*   By: mfujishi <mfujishi@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/06/24 22:16:02 by tkomatsu          #+#    #+#             */
-/*   Updated: 2021/03/11 18:35:16 by tkomatsu         ###   ########.fr       */
+/*   Updated: 2021/05/11 11:49:24 by tkomatsu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,11 +14,11 @@
 
 char	*ft_strrchr(const char *s, int c)
 {
-	int		i;
+	size_t	i;
 	char	*ch;
 
 	ch = (char *)s;
-	i = (int)ft_strlen(s);
+	i = ft_strlen(s);
 	while (i > 0 && s[i] != (char)c)
 		i--;
 	if (ch[i] == (char)c)
